requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
mysql-connector-python
hdbcli
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gunicorn
requests
Flask-Mail
pandas
xlsxwriter
PyJWT